Fix TypeError when exc_info=False
Formatting of a message fails in the `StdlibFormatter` when the
`exc_info` parameter is `False`. The traceback shows that handling of
the boolean is not done correct in the formatting for error type and
message.

```python
Traceback (most recent call last):
  File "lib/python3.8/logging/__init__.py", line 1081, in emit
    msg = self.format(record)
  File "lib/python3.8/logging/__init__.py", line 925, in format
    return fmt.format(record)
  File "ecs_logging/_stdlib.py", line 117, in format
    result = self.format_to_ecs(record)
  File "ecs_logging/_stdlib.py", line 163, in format_to_ecs
    value = extractors[field](record)
  File "ecs_logging/_stdlib.py", line 150, in <lambda>
    if (r.exc_info is not None and r.exc_info[0] is not None)
TypeError: 'bool' object is not subscriptable
```

This replaces the lambda with methods handling `exc_info` being an
boolean or iterable.